Change context ptebase to 64 bits

+++ src/n64/headers/n64/cpu/cop0.h
@@ -59,7 +59,7 @@ struct Index
 struct Context
 {
     u32 bad_vpn2 = 0;
-    u32 pte_base = 0;
+    u64 pte_base = 0;
 };
 
 struct Config {
